feat: change handle fn args, handle result by type
object, number, and boolean uses res.json so that the content type is
set to application/json, while string uses res.send to allow sending
HTML strings.

bigint, symbol, and function throws error (since they are not meant
to be JSON-serializable).

With context object, we can add request helpers in the future without
having to modify the Express/Node req/res object.

BREAKING CHANGE: handle is now incompatible with existing handle
implementations, which put req and res as first and second args for the
handler function.